ggml-cpu : fix linking order
